refactor: split FdbBinding handlers into one file per command
