+++ packages/fdb_helper/lib/src/fdb_binding.dart
@@ -1,51 +1,19 @@
-import 'dart:convert';
+// fdb_binding.dart — thin binding: registration only, zero logic
 import 'dart:developer' as developer;
-import 'dart:io';
-import 'dart:ui' as ui;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
-import 'package:path_provider/path_provider.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
-import 'element_tree_finder.dart';
-import 'gesture_dispatcher.dart';
-import 'hit_test_utils.dart';
-import 'text_input_simulator.dart';
-import 'widget_matcher.dart';
-
-/// Minimum pixel movement required to consider a scroll attempt as non-stalled.
-/// If the scroll position changes by less than this amount after a drag gesture,
-/// the attempt is counted as a stall.
-const double _stallThresholdPixels = 0.5;
-
-/// How long to wait after a drag gesture for scroll physics to settle before
-/// reading [ScrollPosition.pixels], [ScrollPosition.extentAfter], and
-/// [ScrollPosition.extentBefore].
-const Duration _scrollSettleDuration = Duration(milliseconds: 80);
-
-/// How many consecutive edge readings are required before treating the scroll
-/// edge as truly reached. Guards against transient zero readings of
-/// [ScrollPosition.extentAfter] (forward edge) or [ScrollPosition.extentBefore]
-/// (backward edge) that can occur during [ListView.builder] rebuild mid-scroll.
-const int _edgeConfirmationCount = 2;
-
-/// How many consecutive stall readings are required before treating the scroll
-/// as truly stuck and reversing direction. A stall is counted when
-/// [ScrollPosition.pixels] changes by less than [_stallThresholdPixels] after
-/// a drag gesture.
-const int _stallConfirmationCount = 3;
-
-/// Minimum scroll range (in pixels) a Scrollable must have to be considered
-/// a meaningful scroll container. Scrollables with a range at or below this
-/// value (e.g. fully-visible lists with no overflow) are ignored when
-/// selecting the best Scrollable for scroll-to.
-const double _minScrollRangePixels = 0.5;
-
-/// Pixel distance of each drag gesture used by scroll-to.
-/// Matches the default step used by `fdb scroll`, large enough to produce
-/// scroll physics momentum on all platforms.
-const double _scrollStepPixels = 200.0;
+import 'handlers/back_handler.dart';
+import 'handlers/clean_handler.dart';
+import 'handlers/describe_handler.dart';
+import 'handlers/input_handler.dart';
+import 'handlers/screenshot_handler.dart';
+import 'handlers/scroll_handler.dart';
+import 'handlers/scroll_to_handler.dart';
+import 'handlers/shared_prefs_handler.dart';
+import 'handlers/swipe_handler.dart';
+import 'handlers/tap_handler.dart';
 
 /// A custom binding that registers VM service extensions for widget interaction.
 ///
@@ -59,12 +27,16 @@ const double _scrollStepPixels = 200.0;
 ///
 /// This registers the following VM service extensions (in debug and profile mode only):
 /// - `ext.fdb.elements` — list all interactive elements with bounds
+/// - `ext.fdb.describe` — describe the current screen
 /// - `ext.fdb.tap` — tap a widget by key, text, type, or coordinates
 /// - `ext.fdb.longPress` — long-press a widget (same as tap with duration=500ms)
 /// - `ext.fdb.enterText` — enter text into a text field
 /// - `ext.fdb.scroll` — perform a swipe/scroll gesture
 /// - `ext.fdb.scrollTo` — scroll until a target widget becomes visible
+/// - `ext.fdb.swipe` — swipe in a direction
 /// - `ext.fdb.back` — trigger Navigator.maybePop()
+/// - `ext.fdb.clean` — delete app storage directories
+/// - `ext.fdb.sharedPrefs` — read/write shared preferences
 /// - `ext.fdb.screenshot` — capture the Flutter rendering surface as base64 PNG
 class FdbBinding extends WidgetsFlutterBinding {
   FdbBinding._();
@@ -88,28 +60,24 @@ class FdbBinding extends WidgetsFlutterBinding {
   @override
   void initServiceExtensions() {
     super.initServiceExtensions();
-
     if (kReleaseMode) return;
-
-    _registerExtension('ext.fdb.elements', _handleElements);
-    _registerExtension('ext.fdb.describe', _handleDescribe);
-    _registerExtension('ext.fdb.tap', _handleTap);
+    _registerExtension('ext.fdb.elements', handleElements);
+    _registerExtension('ext.fdb.describe', handleDescribe);
+    _registerExtension('ext.fdb.tap', handleTap);
     _registerExtension('ext.fdb.longPress', (method, params) {
-      // Long-press is identical to tap but defaults to 500 ms hold duration.
-      final paramsWithDuration = {
+      return handleTap(method, {
         ...params,
         if (!params.containsKey('duration')) 'duration': '500',
-      };
-      return _handleTap(method, paramsWithDuration);
+      });
     });
-    _registerExtension('ext.fdb.enterText', _handleEnterText);
-    _registerExtension('ext.fdb.scroll', _handleScroll);
-    _registerExtension('ext.fdb.scrollTo', _handleScrollTo);
-    _registerExtension('ext.fdb.swipe', _handleSwipe);
-    _registerExtension('ext.fdb.back', _handleBack);
-    _registerExtension('ext.fdb.clean', _handleClean);
-    _registerExtension('ext.fdb.sharedPrefs', _handleSharedPrefs);
-    _registerExtension('ext.fdb.screenshot', _handleScreenshot);
+    _registerExtension('ext.fdb.enterText', handleEnterText);
+    _registerExtension('ext.fdb.scroll', handleScroll);
+    _registerExtension('ext.fdb.scrollTo', handleScrollTo);
+    _registerExtension('ext.fdb.swipe', handleSwipe);
+    _registerExtension('ext.fdb.back', handleBack);
+    _registerExtension('ext.fdb.clean', handleClean);
+    _registerExtension('ext.fdb.sharedPrefs', handleSharedPrefs);
+    _registerExtension('ext.fdb.screenshot', handleScreenshot);
   }
 
   /// Registers a VM service extension, silently ignoring double-registration
@@ -126,1411 +94,5 @@ class FdbBinding extends WidgetsFlutterBinding {
     } on ArgumentError {
       // Already registered — happens on hot reload. Safe to ignore.
     }
-  }
-
-  // ---------------------------------------------------------------------------
-  // Extension handlers
-  // ---------------------------------------------------------------------------
-
-  Future<developer.ServiceExtensionResponse> _handleElements(
-    String method,
-    Map<String, String> params,
-  ) async {
-    try {
-      final elements = findInteractiveElements();
-      return developer.ServiceExtensionResponse.result(
-        jsonEncode({'status': 'Success', 'elements': elements}),
-      );
-    } catch (e) {
-      return _errorResponse('Failed to list elements: $e');
-    }
-  }
-
-  Future<developer.ServiceExtensionResponse> _handleDescribe(
-    String method,
-    Map<String, String> params,
-  ) async {
-    try {
-      final root = WidgetsBinding.instance.rootElement;
-      if (root == null) {
-        return _errorResponse('No root element available');
-      }
-
-      final view = WidgetsBinding.instance.platformDispatcher.views.first;
-      final screenSize = view.physicalSize / view.devicePixelRatio;
-      final screenRect = Offset.zero & screenSize;
-
-      final interactive = <Map<String, dynamic>>[];
-      final texts = <String>{};
-      String? screenName;
-      String? routeName;
-
-      // Tracks the nearest enclosing Tooltip message while walking the tree.
-      String? currentTooltip;
-
-      void visit(Element element) {
-        final widget = element.widget;
-        final typeName = widget.runtimeType.toString();
-
-        // Skip entire subtrees wrapped in Offstage(offstage: true).
-        // Flutter uses Offstage to hide inactive routes, so this naturally
-        // filters background route content from the describe output.
-        if (typeName == 'Offstage') {
-          try {
-            final offstage = (widget as dynamic).offstage as bool;
-            if (offstage) return;
-          } catch (_) {}
-        }
-
-        // Collect route name from ModalRoute — no guard so the deepest
-        // (topmost visible) route always wins over shallower ancestors.
-        if (element is StatefulElement) {
-          final route = ModalRoute.of(element);
-          if (route != null) {
-            routeName = route.settings.name;
-          }
-        }
-
-        // Collect Scaffold title as screen name by walking AppBar children
-        if (screenName == null && typeName == 'Scaffold') {
-          // Walk children to find AppBar → title Text
-          void findAppBarTitle(Element el) {
-            if (screenName != null) return;
-            final w = el.widget;
-            final wType = w.runtimeType.toString();
-            if (wType == 'AppBar' || wType == 'SliverAppBar') {
-              // Walk AppBar children to find the title Text
-              void findTitle(Element titleEl) {
-                if (screenName != null) return;
-                final tw = titleEl.widget;
-                if (tw is Text) {
-                  final t = tw.data ?? tw.textSpan?.toPlainText();
-                  if (t != null && t.trim().isNotEmpty) {
-                    screenName = t.trim();
-                  }
-                  return;
-                }
-                titleEl.visitChildren(findTitle);
-              }
-
-              el.visitChildren(findTitle);
-              return;
-            }
-            el.visitChildren(findAppBarTitle);
-          }
-
-          element.visitChildren(findAppBarTitle);
-        }
-
-        // Track Tooltip context so inner interactive widgets can inherit it.
-        String? previousTooltip;
-        if (typeName == 'Tooltip') {
-          previousTooltip = currentTooltip;
-          try {
-            final msg = (widget as dynamic).message as String?;
-            if (msg != null && msg.trim().isNotEmpty) {
-              currentTooltip = msg.trim();
-            }
-          } catch (_) {}
-        }
-
-        final renderObject = element.renderObject;
-        if (renderObject is RenderBox && renderObject.hasSize && renderObject.attached) {
-          final offset = renderObject.localToGlobal(Offset.zero);
-          final size = renderObject.size;
-
-          // Skip zero-size or off-screen widgets
-          if (size.isEmpty) {
-            element.visitChildren(visit);
-            if (typeName == 'Tooltip') currentTooltip = previousTooltip;
-            return;
-          }
-          final elementRect = offset & size;
-          if (!screenRect.overlaps(elementRect)) {
-            element.visitChildren(visit);
-            if (typeName == 'Tooltip') currentTooltip = previousTooltip;
-            return;
-          }
-
-          // Collect visible text
-          if (widget is Text) {
-            final text = widget.data ?? widget.textSpan?.toPlainText();
-            if (text != null && text.trim().isNotEmpty) {
-              texts.add(text.trim());
-            }
-          } else if (widget is RichText) {
-            final text = widget.text.toPlainText().trim();
-            if (text.isNotEmpty) texts.add(text);
-          } else if (widget is EditableText) {
-            final text = widget.controller.text.trim();
-            if (text.isNotEmpty) texts.add(text);
-          }
-
-          // Collect interactive widgets
-          if (_isDescribeInteractiveWidget(typeName)) {
-            // Skip widgets obscured by a foreground route (e.g. GoRouter shell
-            // routes where background pages aren't wrapped in Offstage).
-            // If the hit test at this widget's center doesn't reach its render
-            // object, a user can't interact with it — omit it and its children.
-            if (!isElementHittable(element)) {
-              if (typeName == 'Tooltip') currentTooltip = previousTooltip;
-              return;
-            }
-
-            final key = widget.key is ValueKey<String> ? (widget.key as ValueKey<String>).value : null;
-            final visibleText = _extractDescribeText(element, tooltipHint: currentTooltip);
-            final gestures = _extractGestures(widget, typeName);
-            interactive.add({
-              'type': typeName,
-              'key': key,
-              'text': visibleText,
-              'x': offset.dx + size.width / 2,
-              'y': offset.dy + size.height / 2,
-              if (gestures != null) 'gestures': gestures,
-            });
-
-            // Still collect text from children for the TEXT section even
-            // though we stop recursing for interactive-widget purposes.
-            void collectText(Element el) {
-              final w = el.widget;
-              if (w is Text) {
-                final t = w.data ?? w.textSpan?.toPlainText();
-                if (t != null && t.trim().isNotEmpty) texts.add(t.trim());
-              } else if (w is RichText) {
-                final t = w.text.toPlainText().trim();
-                if (t.isNotEmpty) texts.add(t);
-              }
-              el.visitChildren(collectText);
-            }
-
-            element.visitChildren(collectText);
-
-            // Don't recurse into interactive widgets to avoid duplicates
-            if (typeName == 'Tooltip') currentTooltip = previousTooltip;
-            return;
-          }
-        }
-
-        element.visitChildren(visit);
-        if (typeName == 'Tooltip') currentTooltip = previousTooltip;
-      }
-
-      root.visitChildren(visit);
-
-      // Remove interactive entries that provide no useful identity signal:
-      // no text, no key, and no interesting gestures (drag/pan/longPress/scale).
-      // A bare GestureDetector(tap) with no text/key is almost always an icon
-      // or structural element that the agent cannot meaningfully reference.
-      const interestingGestures = {
-        'horizontalDrag',
-        'verticalDrag',
-        'pan',
-        'longPress',
-        'scale',
-        'doubleTap',
-        'forcePress',
-      };
-      interactive.removeWhere((entry) {
-        final text = entry['text'] as String?;
-        final key = entry['key'] as String?;
-        final gestures = (entry['gestures'] as List<dynamic>?)?.cast<String>() ?? [];
-        // Text consisting only of Unicode PUA icon codepoints (U+E000-U+F8FF)
-        // is not meaningful — treat it as empty.
-        final hasText = text != null && text.trim().isNotEmpty && text.runes.any((r) => r < 0xE000 || r > 0xF8FF);
-        final hasKey = key != null;
-        final hasInterestingGesture = gestures.any((g) => interestingGestures.contains(g));
-        return !hasText && !hasKey && !hasInterestingGesture;
-      });
-
-      // Sort interactive elements top-to-bottom, left-to-right
-      interactive.sort((a, b) {
-        final yA = (a['y'] as double);
-        final yB = (b['y'] as double);
-        if ((yA - yB).abs() > 10) return yA.compareTo(yB);
-        return (a['x'] as double).compareTo(b['x'] as double);
-      });
-
-      // Assign sequential refs
-      for (var i = 0; i < interactive.length; i++) {
-        interactive[i]['ref'] = i + 1;
-      }
-
-      return developer.ServiceExtensionResponse.result(
-        jsonEncode({
-          'status': 'Success',
-          'screen': screenName,
-          'route': routeName,
-          'interactive': interactive,
-          'texts': texts.toList(),
-        }),
-      );
-    } catch (e) {
-      return _errorResponse('describe failed: $e');
-    }
-  }
-
-  /// Returns true for widget types that are interactive and should appear in
-  /// the describe output. Extends [_isInteractiveWidget] with additional
-  /// tappable container types.
-  bool _isDescribeInteractiveWidget(String typeName) => const {
-        'Checkbox',
-        'CheckboxListTile',
-        'DropdownButton',
-        'ElevatedButton',
-        'FilledButton',
-        'FloatingActionButton',
-        'GestureDetector',
-        'IconButton',
-        'InkWell',
-        'ListTile',
-        'OutlinedButton',
-        'PopupMenuButton',
-        'Radio',
-        'RadioListTile',
-        'Slider',
-        'Switch',
-        'SwitchListTile',
-        'Tab',
-        'TextButton',
-        'TextField',
-        'TextFormField',
-      }.contains(typeName);
-
-  /// Extracts visible text from an interactive element by walking its children.
-  ///
-  /// Collects ALL text fragments (Text, RichText, EditableText), Tooltip
-  /// messages, and Icon semantic labels, joining them with " · ".
-  ///
-  /// [tooltipHint] is the nearest enclosing Tooltip message discovered while
-  /// walking the tree top-down; it is used as a fallback when no other text
-  /// is found inside the widget.
-  String? _extractDescribeText(
-    Element element, {
-    String? tooltipHint,
-  }) {
-    // Check the widget itself first (leaf cases).
-    final widget = element.widget;
-    if (widget is Text) return widget.data ?? widget.textSpan?.toPlainText();
-    if (widget is RichText) return widget.text.toPlainText();
-    if (widget is EditableText) return widget.controller.text;
-
-    // Collect all text fragments from the subtree.
-    final fragments = <String>[];
-
-    // Returns true if the string has at least one non-PUA character.
-    bool hasVisibleText(String s) => s.trim().isNotEmpty && s.runes.any((r) => r < 0xE000 || r > 0xF8FF);
-
-    void findTextAndIcons(Element el) {
-      final w = el.widget;
-
-      if (w is Text) {
-        final t = w.data ?? w.textSpan?.toPlainText();
-        if (t != null && hasVisibleText(t)) fragments.add(t.trim());
-        return; // Don't recurse into Text children
-      }
-      if (w is RichText) {
-        final t = w.text.toPlainText().trim();
-        if (hasVisibleText(t)) fragments.add(t);
-        return;
-      }
-      if (w is EditableText) {
-        final t = w.controller.text.trim();
-        if (hasVisibleText(t)) fragments.add(t);
-        return;
-      }
-
-      final wType = w.runtimeType.toString();
-
-      // Extract Tooltip message (inner Tooltip, distinct from the ancestor one).
-      if (wType == 'Tooltip') {
-        try {
-          final message = (w as dynamic).message as String?;
-          if (message != null && message.trim().isNotEmpty) {
-            fragments.add('[${message.trim()}]');
-          }
-        } catch (_) {}
-      }
-
-      // Extract Icon semantic label.
-      if (wType == 'Icon') {
-        try {
-          final label = (w as dynamic).semanticLabel as String?;
-          if (label != null && label.trim().isNotEmpty) {
-            fragments.add('[icon: ${label.trim()}]');
-          }
-        } catch (_) {}
-      }
-
-      el.visitChildren(findTextAndIcons);
-    }
-
-    element.visitChildren(findTextAndIcons);
-
-    // Fall back to the ancestor Tooltip hint when the widget has no own text.
-    if (fragments.isEmpty && tooltipHint != null) {
-      return '[${tooltipHint.trim()}]';
-    }
-
-    if (fragments.isEmpty) return null;
-    final cleaned = fragments.where((f) => f.trim().isNotEmpty).toList();
-    if (cleaned.isEmpty) return null;
-    return cleaned.join(' · ');
-  }
-
-  /// Extracts the list of registered gestures for a widget.
-  ///
-  /// For [GestureDetector] and [InkWell], checks which callback properties
-  /// are non-null using dynamic access. Returns a short list like
-  /// `["tap", "longPress", "horizontalDrag"]` or null if not applicable.
-  List<String>? _extractGestures(Widget widget, String typeName) {
-    if (typeName != 'GestureDetector' && typeName != 'InkWell') return null;
-
-    final gestures = <String>[];
-    try {
-      final w = widget as dynamic;
-      // Tap gestures
-      if (typeName == 'GestureDetector') {
-        if (w.onTap != null) gestures.add('tap');
-        if (w.onDoubleTap != null) gestures.add('doubleTap');
-        if (w.onLongPress != null) gestures.add('longPress');
-        if (w.onVerticalDragStart != null || w.onVerticalDragUpdate != null || w.onVerticalDragEnd != null) {
-          gestures.add('verticalDrag');
-        }
-        if (w.onHorizontalDragStart != null || w.onHorizontalDragUpdate != null || w.onHorizontalDragEnd != null) {
-          gestures.add('horizontalDrag');
-        }
-        if (w.onPanStart != null || w.onPanUpdate != null || w.onPanEnd != null) {
-          gestures.add('pan');
-        }
-        if (w.onScaleStart != null || w.onScaleUpdate != null || w.onScaleEnd != null) {
-          gestures.add('scale');
-        }
-        if (w.onForcePressStart != null || w.onForcePressPeak != null) {
-          gestures.add('forcePress');
-        }
-      } else if (typeName == 'InkWell') {
-        if (w.onTap != null) gestures.add('tap');
-        if (w.onDoubleTap != null) gestures.add('doubleTap');
-        if (w.onLongPress != null) gestures.add('longPress');
-      }
-    } catch (_) {
-      // Dynamic access failed — widget API changed; return what we have.
-    }
-    return gestures.isEmpty ? null : gestures;
-  }
-
-  Future<developer.ServiceExtensionResponse> _handleTap(
-    String method,
-    Map<String, String> params,
-  ) async {
-    try {
-      final rawDuration = params['duration'];
-      final durationMs = rawDuration != null ? int.tryParse(rawDuration) : null;
-      if (rawDuration != null && durationMs == null) {
-        return _errorResponse('Invalid duration value: $rawDuration');
-      }
-      final holdDuration = durationMs != null ? Duration(milliseconds: durationMs) : const Duration(milliseconds: 10);
-
-      final matcher = WidgetMatcher.fromParams(params);
-
-      if (matcher is CoordinatesMatcher) {
-        await dispatchTap(matcher.offset, holdDuration: holdDuration);
-        return developer.ServiceExtensionResponse.result(
-          jsonEncode({'status': 'Success', 'x': matcher.x, 'y': matcher.y}),
-        );
-      }
-
-      final (:element, :matchCount) = findHittableElement(matcher);
-      if (element == null) {
-        if (matchCount > 1) {
-          return _errorResponse(
-            'Found $matchCount elements matching the selector. '
-            'Use --index to specify which one (0-based).',
-          );
-        }
-        return _errorResponse('No hittable element found for matcher');
-      }
-
-      final renderObject = element.renderObject;
-      if (renderObject is! RenderBox) {
-        return _errorResponse('Element has no RenderBox');
-      }
-
-      final center = renderObject.size.center(Offset.zero);
-      final globalCenter = renderObject.localToGlobal(center);
-      await dispatchTap(globalCenter, holdDuration: holdDuration);
-
-      final widgetType = element.widget.runtimeType.toString();
-      return developer.ServiceExtensionResponse.result(
-        jsonEncode({
-          'status': 'Success',
-          'widgetType': widgetType,
-          'x': globalCenter.dx,
-          'y': globalCenter.dy,
-        }),
-      );
-    } on ArgumentError catch (e) {
-      return _errorResponse(e.message.toString());
-    } catch (e) {
-      return _errorResponse('Tap failed: $e');
-    }
-  }
-
-  Future<developer.ServiceExtensionResponse> _handleEnterText(
-    String method,
-    Map<String, String> params,
-  ) async {
-    try {
-      final input = params['input'];
-      if (input == null) {
-        return _errorResponse('Missing required param: input');
-      }
-
-      final matcher = WidgetMatcher.fromParams(params);
-
-      if (matcher is FocusedMatcher) {
-        // Type into the currently focused element.
-        final focusContext = FocusManager.instance.primaryFocus?.context;
-        if (focusContext == null) {
-          return _errorResponse('No focused element found');
-        }
-
-        // Walk descendants to find the nearest EditableText element.
-        // EditableText is a child of TextField, not a parent.
-        Element? editableElement;
-
-        void findEditableText(Element el) {
-          if (editableElement != null) return;
-          if (el.widget is EditableText) {
-            editableElement = el;
-            return;
-          }
-          el.visitChildElements(findEditableText);
-        }
-
-        if (focusContext is Element) {
-          findEditableText(focusContext);
-        }
-
-        if (editableElement == null) {
-          return _errorResponse(
-            'Focused element is not an editable text field',
-          );
-        }
-
-        await enterText(editableElement!, input);
-
-        return developer.ServiceExtensionResponse.result(
-          jsonEncode({
-            'status': 'Success',
-            'input': input,
-            'widgetType': editableElement!.widget.runtimeType.toString(),
-          }),
-        );
-      }
-
-      final (:element, :matchCount) = findHittableElement(matcher);
-      if (element == null) {
-        if (matchCount > 1) {
-          return _errorResponse(
-            'Found $matchCount elements matching the selector. '
-            'Use --index to specify which one (0-based).',
-          );
-        }
-        return _errorResponse('No hittable element found for matcher');
-      }
-
-      await enterText(element, input);
-
-      return developer.ServiceExtensionResponse.result(
-        jsonEncode({
-          'status': 'Success',
-          'input': input,
-          'widgetType': element.widget.runtimeType.toString(),
-        }),
-      );
-    } on ArgumentError catch (e) {
-      return _errorResponse(e.message.toString());
-    } catch (e) {
-      return _errorResponse('enterText failed: $e');
-    }
-  }
-
-  Future<developer.ServiceExtensionResponse> _handleScroll(
-    String method,
-    Map<String, String> params,
-  ) async {
-    try {
-      // Support two modes:
-      // 1. direction + distance (+ optional at=x,y)
-      // 2. raw startX/startY/endX/endY
-
-      final direction = params['direction'];
-
-      double startX;
-      double startY;
-      double endX;
-      double endY;
-
-      if (direction != null) {
-        final distance = double.tryParse(params['distance'] ?? '') ?? 200.0;
-
-        // Default start: center of screen
-        final view = WidgetsBinding.instance.platformDispatcher.views.first;
-        final screenSize = view.physicalSize / view.devicePixelRatio;
-        var centerX = screenSize.width / 2;
-        var centerY = screenSize.height / 2;
-
-        // Override with --at x,y
-        final at = params['at'];
-        if (at != null) {
-          final parts = at.split(',');
-          final atX = parts.length == 2 ? double.tryParse(parts[0]) : null;
-          final atY = parts.length == 2 ? double.tryParse(parts[1]) : null;
-          if (atX == null || atY == null) {
-            return _errorResponse(
-              'Invalid --at value: "$at". Expected format: x,y (e.g. 200,400).',
-            );
-          }
-          centerX = atX;
-          centerY = atY;
-        }
-
-        startX = centerX;
-        startY = centerY;
-
-        switch (direction) {
-          case 'up':
-            // Finger moves down → content scrolls up (reveals content above).
-            endX = startX;
-            endY = startY + distance;
-          case 'down':
-            // Finger moves up → content scrolls down (reveals content below).
-            endX = startX;
-            endY = startY - distance;
-          case 'left':
-            // Finger moves right → content scrolls left.
-            endX = startX + distance;
-            endY = startY;
-          case 'right':
-            // Finger moves left → content scrolls right.
-            endX = startX - distance;
-            endY = startY;
-          default:
-            return _errorResponse(
-              'Invalid direction: $direction. Use up, down, left, or right.',
-            );
-        }
-      } else {
-        // Raw coordinates mode
-        final sx = double.tryParse(params['startX'] ?? '');
-        final sy = double.tryParse(params['startY'] ?? '');
-        final ex = double.tryParse(params['endX'] ?? '');
-        final ey = double.tryParse(params['endY'] ?? '');
-
-        if (sx == null || sy == null || ex == null || ey == null) {
-          return _errorResponse(
-            'Provide direction (up/down/left/right) or '
-            'startX, startY, endX, endY',
-          );
-        }
-
-        startX = sx;
-        startY = sy;
-        endX = ex;
-        endY = ey;
-      }
-
-      await dispatchScroll(
-        start: Offset(startX, startY),
-        end: Offset(endX, endY),
-      );
-
-      return developer.ServiceExtensionResponse.result(
-        jsonEncode({
-          'status': 'Success',
-          'startX': startX,
-          'startY': startY,
-          'endX': endX,
-          'endY': endY,
-        }),
-      );
-    } catch (e) {
-      return _errorResponse('Scroll failed: $e');
-    }
-  }
-
-  Future<developer.ServiceExtensionResponse> _handleScrollTo(
-    String method,
-    Map<String, String> params,
-  ) async {
-    try {
-      final matcher = WidgetMatcher.fromParams(params);
-
-      if (matcher is FocusedMatcher) {
-        return _errorResponse('scroll-to does not support --focused');
-      }
-
-      // Check if the target widget is already visible before requiring a Scrollable.
-      final earlyResult = findHittableElement(matcher);
-      if (earlyResult.element != null) {
-        // Use the raw matched element for ensureVisible so that Text widgets
-        // inside non-interactive containers (e.g. Column) are positioned
-        // correctly, rather than centering the entire ancestor container.
-        final ensureVisibleTarget =
-            findScrollTargetElement(matcher) ?? earlyResult.element!;
-        final earlyRenderObject = ensureVisibleTarget.renderObject;
-        if (earlyRenderObject is RenderBox) {
-          // Look up the scroll position so we can stop any in-progress
-          // ballistic animation before calling ensureVisible, matching the
-          // behaviour of the main loop and post-loop paths.
-          ScrollPosition? earlyPosition;
-          ensureVisibleTarget.visitAncestorElements((ancestor) {
-            if (ancestor.widget is Scrollable) {
-              earlyPosition = _tryGetScrollPosition(ancestor);
-              return false;
-            }
-            return true;
-          });
-          // alignment: 0.0 (Flutter default) avoids unnecessary scrolling when
-          // the widget is already fully visible on screen.
-          return await _ensureVisibleAndReport(
-            ensureVisibleTarget: ensureVisibleTarget,
-            fallbackRenderObject: earlyRenderObject,
-            position: earlyPosition,
-            alignment: 0.0,
-          );
-        }
-      }
-      if (earlyResult.matchCount > 1) {
-        return _errorResponse(
-          'Found ${earlyResult.matchCount} elements matching the selector. '
-          'Use --index to specify which one (0-based).',
-        );
-      }
-
-      final scrollableElement = _findBestScrollable(matcher);
-      if (scrollableElement == null) {
-        return _errorResponse('No Scrollable found in the widget tree');
-      }
-
-      final position = _tryGetScrollPosition(scrollableElement);
-      if (position == null) {
-        return _errorResponse('Could not read ScrollPosition from Scrollable');
-      }
-
-      final scrollableWidget = scrollableElement.widget as Scrollable;
-      final axisDirection = scrollableWidget.axisDirection;
-
-      // Initial finger movement direction to scroll toward the end of the list.
-      // The gesture direction is opposite to the content movement direction:
-      //   down-axis: drag finger up (dy < 0) → content scrolls down
-      //   up-axis:   drag finger down (dy > 0) → content scrolls up
-      //   right-axis: drag finger left (dx < 0) → content scrolls right
-      //   left-axis:  drag finger right (dx > 0) → content scrolls left
-      // _scrollStepPixels matches the default used by fdb scroll, which is
-      // large enough to produce scroll physics momentum on all platforms.
-      var moveStep = switch (axisDirection) {
-        AxisDirection.down => const Offset(0, -_scrollStepPixels),
-        AxisDirection.up => const Offset(0, _scrollStepPixels),
-        AxisDirection.right => const Offset(-_scrollStepPixels, 0),
-        AxisDirection.left => const Offset(_scrollStepPixels, 0),
-      };
-
-      final maxAttempts = _calculateMaxAttempts(position);
-      var reversedOnce = false;
-      var stallCount = 0;
-      var edgeCount = 0;
-      var attempt = 0;
-
-      while (attempt < maxAttempts) {
-        // Check if target is now in tree and hittable.
-        final (:element, :matchCount) = findHittableElement(matcher);
-        if (element != null) {
-          final renderObject = element.renderObject;
-          if (renderObject is RenderBox) {
-            // Stop any ongoing ballistic scroll animation by jumping to the
-            // current position, then precisely position the target in the
-            // viewport using ensureVisible.
-            //
-            // Use the raw matched element (e.g. the Text widget itself) for
-            // ensureVisible so that the scroll offset is computed relative to
-            // the actual target, not an ancestor container like Column.
-            // alignment: 0.5 centers the item in the viewport, which is more
-            // reliable across platforms and avoids off-by-one edge cases.
-            final ensureVisibleTarget =
-                findScrollTargetElement(matcher) ?? element;
-            return await _ensureVisibleAndReport(
-              ensureVisibleTarget: ensureVisibleTarget,
-              fallbackRenderObject: renderObject,
-              position: position,
-              alignment: 0.5,
-            );
-          }
-        }
-        if (matchCount > 1) {
-          return _errorResponse(
-            'Found $matchCount elements matching the selector. '
-            'Use --index to specify which one (0-based).',
-          );
-        }
-
-        // Perform drag gesture on the Scrollable's center.
-        final scrollableRenderObject = scrollableElement.renderObject;
-        if (scrollableRenderObject is! RenderBox) break;
-
-        final scrollableCenter =
-            scrollableRenderObject.size.center(Offset.zero);
-        final scrollableGlobalCenter =
-            scrollableRenderObject.localToGlobal(scrollableCenter);
-
-        final currentPixels = position.pixels;
-
-        await dispatchScroll(
-          start: scrollableGlobalCenter,
-          end: scrollableGlobalCenter + moveStep,
-        );
-
-        // Wait for scroll physics to settle before reading position.
-        // ListView.builder needs a frame to build newly visible items and
-        // update extentAfter accurately after each drag.
-        await Future<void>.delayed(_scrollSettleDuration);
-
-        // Stall detection: if position didn't change, reverse or give up.
-        final newPixels = position.pixels;
-        if ((newPixels - currentPixels).abs() < _stallThresholdPixels) {
-          stallCount++;
-          if (stallCount >= _stallConfirmationCount) {
-            if (!reversedOnce) {
-              moveStep = -moveStep;
-              reversedOnce = true;
-              stallCount = 0;
-              edgeCount = 0;
-              attempt = 0; // Reset full budget for the reversed direction.
-              continue;
-            } else {
-              break;
-            }
-          }
-        } else {
-          stallCount = 0;
-        }
-
-        // Detect scroll edge to decide whether to reverse or give up.
-        // Require _edgeConfirmationCount consecutive readings to avoid acting
-        // on transient extentAfter==0 that can occur during ListView.builder
-        // rebuilds mid-scroll.
-        final scrollingForward = switch (axisDirection) {
-          AxisDirection.down => moveStep.dy < 0,
-          AxisDirection.up => moveStep.dy > 0,
-          AxisDirection.right => moveStep.dx < 0,
-          AxisDirection.left => moveStep.dx > 0,
-        };
-        final atCurrentEdge = scrollingForward
-            ? position.extentAfter <= 0
-            : position.extentBefore <= 0;
-
-        if (atCurrentEdge) {
-          edgeCount++;
-          if (edgeCount >= _edgeConfirmationCount) {
-            if (!reversedOnce) {
-              moveStep = -moveStep;
-              reversedOnce = true;
-              stallCount = 0;
-              edgeCount = 0;
-              attempt = 0; // Reset full budget for the reversed direction.
-              continue;
-            } else {
-              break; // Already reversed once — searched entire list, give up.
-            }
-          }
-        } else {
-          edgeCount = 0;
-        }
-
-        attempt++;
-      }
-
-      // Final check after loop exhaustion.
-      final (:element, :matchCount) = findHittableElement(matcher);
-      if (element != null) {
-        final renderObject = element.renderObject;
-        if (renderObject is RenderBox) {
-          // Stop any ongoing ballistic scroll animation by jumping to the
-          // current position, then precisely position the target in the
-          // viewport using ensureVisible.
-          final ensureVisibleTarget =
-              findScrollTargetElement(matcher) ?? element;
-          return await _ensureVisibleAndReport(
-            ensureVisibleTarget: ensureVisibleTarget,
-            fallbackRenderObject: renderObject,
-            position: position,
-            alignment: 0.5,
-          );
-        }
-      }
-      if (matchCount > 1) {
-        return _errorResponse(
-          'Found $matchCount elements matching the selector. '
-          'Use --index to specify which one (0-based).',
-        );
-      }
-
-      return _errorResponse(
-        'Widget not found after scrolling through the list',
-      );
-    } on ArgumentError catch (e) {
-      return _errorResponse(e.message.toString());
-    } catch (e) {
-      return _errorResponse('scrollTo failed: $e');
-    }
-  }
-
-  /// Stops any in-progress ballistic animation, calls [Scrollable.ensureVisible]
-  /// on [ensureVisibleTarget], waits for the frame to settle, then returns a
-  /// [developer.ServiceExtensionResponse] with the widget's global coordinates.
-  ///
-  /// If [position] is provided, [ScrollPosition.jumpTo] is called first to
-  /// cancel any ongoing ballistic scroll animation before [ensureVisible] runs.
-  ///
-  /// After [ensureVisible] completes, the render object is re-read from
-  /// [ensureVisibleTarget] so that the reported coordinates reflect the widget's
-  /// new position. If the re-read render object is not a [RenderBox], the
-  /// provided [fallbackRenderObject] is used instead.
-  ///
-  /// [alignment] is forwarded to [Scrollable.ensureVisible]. Use `0.0` (Flutter
-  /// default) when the widget is already on screen to avoid unnecessary
-  /// scrolling; use `0.5` to center the widget after scrolling it into view.
-  Future<developer.ServiceExtensionResponse> _ensureVisibleAndReport({
-    required Element ensureVisibleTarget,
-    required RenderBox fallbackRenderObject,
-    ScrollPosition? position,
-    required double alignment,
-  }) async {
-    if (position != null) {
-      // jumpTo schedules a frame as a side-effect, so endOfFrame will complete.
-      position.jumpTo(position.pixels);
-    } else {
-      // No jumpTo means no frame is scheduled. Explicitly schedule one so that
-      // the endOfFrame future below does not hang when the app is idle.
-      WidgetsBinding.instance.scheduleFrame();
-    }
-    await WidgetsBinding.instance.endOfFrame;
-    await Scrollable.ensureVisible(
-      ensureVisibleTarget,
-      alignment: alignment,
-      duration: Duration.zero,
-    );
-    // scheduleFrame ensures endOfFrame completes even when ensureVisible does
-    // nothing (widget already at correct position, app otherwise idle).
-    WidgetsBinding.instance.scheduleFrame();
-    await WidgetsBinding.instance.endOfFrame;
-    final targetRenderObject = ensureVisibleTarget.renderObject;
-    final reportRenderObject = targetRenderObject is RenderBox
-        ? targetRenderObject
-        : fallbackRenderObject;
-    final center = reportRenderObject.size.center(Offset.zero);
-    final globalCenter = reportRenderObject.localToGlobal(center);
-    final widgetType = ensureVisibleTarget.widget.runtimeType.toString();
-    return developer.ServiceExtensionResponse.result(
-      jsonEncode({
-        'status': 'Success',
-        'widgetType': widgetType,
-        'x': globalCenter.dx,
-        'y': globalCenter.dy,
-      }),
-    );
-  }
-
-  /// Finds the best [Scrollable] element to use for scroll-to.
-  ///
-  /// If the target widget is already in the tree, walks its ancestors to find
-  /// the nearest enclosing [Scrollable]. Otherwise scans the full tree,
-  /// preferring the **last** (last in depth-first traversal order, i.e. the
-  /// most recently pushed route) [Scrollable] with a non-zero scroll range.
-  ///
-  /// Flutter's Navigator wraps inactive (background) routes in
-  /// [Offstage], so their subtrees are skipped during traversal.
-  /// The active route's Scrollable therefore appears last in a
-  /// depth-first traversal. Picking the last one ensures we target
-  /// the active route rather than a background route.
-  Element? _findBestScrollable(WidgetMatcher matcher) {
-    // Try to find target first — if in tree, walk ancestors.
-    final root = WidgetsBinding.instance.rootElement;
-    if (root == null) return null;
-
-    Element? targetElement;
-    void findTarget(Element el) {
-      if (targetElement != null) return;
-      if (matcher.matches(el, extractText: extractWidgetText)) {
-        targetElement = el;
-        return;
-      }
-      el.visitChildren(findTarget);
-    }
-
-    root.visitChildren(findTarget);
-
-    if (targetElement != null) {
-      Element? ancestor;
-      targetElement!.visitAncestorElements((el) {
-        if (el.widget is Scrollable) {
-          ancestor = el;
-          return false; // stop walking
-        }
-        return true; // keep walking
-      });
-      if (ancestor != null) return ancestor;
-    }
-
-    // Fallback: scan full tree for best Scrollable.
-    // Collect ALL Scrollables with a non-zero range and return the LAST one
-    // found in depth-first order. Because Flutter's Navigator appends newer
-    // routes after older ones in the element tree, the last Scrollable with
-    // range corresponds to the currently active route.
-    Element? fallback;
-    Element? lastWithRange;
-
-    void visitForScrollable(Element el) {
-      // Skip Offstage subtrees (inactive Navigator routes).
-      if (el.widget.runtimeType.toString() == 'Offstage') {
-        try {
-          final offstage = (el.widget as dynamic).offstage as bool;
-          if (offstage) return;
-        } catch (_) {}
-      }
-      if (el.widget is Scrollable) {
-        fallback ??= el;
-        final scrollPosition = _tryGetScrollPosition(el);
-        if (scrollPosition != null) {
-          final range =
-              (scrollPosition.maxScrollExtent - scrollPosition.minScrollExtent)
-                  .abs();
-          if (range > _minScrollRangePixels) {
-            lastWithRange = el; // keep updating — we want the LAST one
-          }
-        }
-      }
-      el.visitChildren(visitForScrollable);
-    }
-
-    root.visitChildren(visitForScrollable);
-    return lastWithRange ?? fallback;
-  }
-
-  /// Extracts the [ScrollPosition] from a [Scrollable] element's state.
-  ScrollPosition? _tryGetScrollPosition(Element scrollableElement) {
-    try {
-      final state =
-          (scrollableElement as StatefulElement).state as ScrollableState;
-      return state.position;
-    } catch (_) {
-      return null;
-    }
-  }
-
-  /// Calculates the maximum number of drag attempts per direction for scroll-to.
-  ///
-  /// This is the budget for one scroll direction. Because the algorithm resets
-  /// the attempt counter to 0 when reversing direction, the total maximum
-  /// number of drag attempts across both directions is `2 * maxAttempts`.
-  ///
-  /// For finite lists, uses `ceil(extent / step size) * 2 + buffer`, clamped
-  /// between `1` and `200`. For infinite lists (e.g. infinite scroll), falls
-  /// back to 50.
-  int _calculateMaxAttempts(ScrollPosition position) {
-    final extent = (position.maxScrollExtent - position.minScrollExtent).abs();
-    if (!extent.isFinite) return 50;
-    return ((extent / _scrollStepPixels).ceil() * 2 + 20).clamp(1, 200);
-  }
-
-  Future<developer.ServiceExtensionResponse> _handleSwipe(
-    String method,
-    Map<String, String> params,
-  ) async {
-    try {
-      final direction = params['direction'];
-      if (direction == null) {
-        return _errorResponse('Missing required param: direction');
-      }
-
-      final view = WidgetsBinding.instance.platformDispatcher.views.first;
-      final screenSize = view.physicalSize / view.devicePixelRatio;
-
-      double startX;
-      double startY;
-      double swipeDistance;
-
-      // Check if a widget selector was provided.
-      final hasSelector = params.containsKey('key') || params.containsKey('text') || params.containsKey('type');
-
-      if (hasSelector) {
-        // Widget-targeted swipe: find the widget's bounds and compute
-        // start = widget center, distance = 60% of widget dimension.
-        final matcher = WidgetMatcher.fromParams(params);
-        final (:element, :matchCount) = findHittableElement(matcher);
-
-        if (element == null) {
-          if (matchCount > 1) {
-            return _errorResponse(
-              'Found $matchCount elements matching the selector. '
-              'Use --index to specify which one (0-based).',
-            );
-          }
-          return _errorResponse('No hittable element found for matcher');
-        }
-
-        final renderObject = element.renderObject;
-        if (renderObject is! RenderBox) {
-          return _errorResponse('Element has no RenderBox');
-        }
-
-        final center = renderObject.size.center(Offset.zero);
-        final globalCenter = renderObject.localToGlobal(center);
-        startX = globalCenter.dx;
-        startY = globalCenter.dy;
-
-        // Smart default: 60% of widget width for left/right, height for up/down.
-        final rawDistance = params['distance'];
-        if (rawDistance != null) {
-          swipeDistance = double.tryParse(rawDistance) ?? 200.0;
-        } else {
-          swipeDistance = switch (direction) {
-            'left' || 'right' => renderObject.size.width * 0.6,
-            'up' || 'down' => renderObject.size.height * 0.6,
-            _ => 200.0,
-          };
-        }
-      } else {
-        // Fallback: screen center (or --at override) + fixed distance.
-        var centerX = screenSize.width / 2;
-        var centerY = screenSize.height / 2;
-
-        final at = params['at'];
-        if (at != null) {
-          final parts = at.split(',');
-          final atX = parts.length == 2 ? double.tryParse(parts[0]) : null;
-          final atY = parts.length == 2 ? double.tryParse(parts[1]) : null;
-          if (atX == null || atY == null) {
-            return _errorResponse(
-              'Invalid --at value: "$at". Expected format: x,y (e.g. 200,400).',
-            );
-          }
-          centerX = atX;
-          centerY = atY;
-        }
-
-        startX = centerX;
-        startY = centerY;
-        swipeDistance = double.tryParse(params['distance'] ?? '') ?? 200.0;
-      }
-
-      double endX;
-      double endY;
-
-      switch (direction) {
-        case 'left':
-          // Finger moves left → content/page moves left (next page).
-          endX = startX - swipeDistance;
-          endY = startY;
-        case 'right':
-          // Finger moves right → content/page moves right (previous page).
-          endX = startX + swipeDistance;
-          endY = startY;
-        case 'up':
-          // Finger moves up → content scrolls up.
-          endX = startX;
-          endY = startY - swipeDistance;
-        case 'down':
-          // Finger moves down → content scrolls down.
-          endX = startX;
-          endY = startY + swipeDistance;
-        default:
-          return _errorResponse(
-            'Invalid direction: $direction. Use up, down, left, or right.',
-          );
-      }
-
-      await dispatchScroll(
-        start: Offset(startX, startY),
-        end: Offset(endX, endY),
-      );
-
-      return developer.ServiceExtensionResponse.result(
-        jsonEncode({
-          'status': 'Success',
-          'direction': direction,
-          'distance': swipeDistance,
-          'startX': startX,
-          'startY': startY,
-          'endX': endX,
-          'endY': endY,
-        }),
-      );
-    } on ArgumentError catch (e) {
-      return _errorResponse(e.message.toString());
-    } catch (e) {
-      return _errorResponse('Swipe failed: $e');
-    }
-  }
-
-  Future<developer.ServiceExtensionResponse> _handleBack(
-    String method,
-    Map<String, String> params,
-  ) async {
-    try {
-      final rootElement = WidgetsBinding.instance.rootElement;
-      if (rootElement == null) {
-        return _errorResponse('No root element available');
-      }
-      // Walk DOWN the tree to find the first NavigatorState.
-      // Navigator.maybeOf(rootElement) walks UP from rootElement, but rootElement
-      // is already the top of the tree — there is nothing above it to find.
-      NavigatorState? navigator;
-      void findNavigator(Element element) {
-        if (navigator != null) return;
-        if (element is StatefulElement && element.state is NavigatorState) {
-          navigator = element.state as NavigatorState;
-          return;
-        }
-        element.visitChildElements(findNavigator);
-      }
-
-      rootElement.visitChildElements(findNavigator);
-      if (navigator == null) {
-        return _errorResponse('No Navigator found');
-      }
-      final popped = await navigator!.maybePop();
-      return developer.ServiceExtensionResponse.result(
-        jsonEncode({'status': 'Success', 'popped': popped}),
-      );
-    } catch (e) {
-      return _errorResponse('Back failed: $e');
-    }
-  }
-
-  /// Handles `ext.fdb.sharedPrefs`.
-  ///
-  /// Params:
-  ///   action — `get` | `getAll` | `set` | `remove` | `clear`
-  ///   key    — required for get / set / remove
-  ///   value  — required for set (always a string; type param determines cast)
-  ///   type   — for set: `string` (default) | `bool` | `int` | `double`
-  Future<developer.ServiceExtensionResponse> _handleSharedPrefs(
-    String method,
-    Map<String, String> params,
-  ) async {
-    try {
-      final prefs = await SharedPreferences.getInstance();
-      final action = params['action'] ?? '';
-
-      switch (action) {
-        case 'getAll':
-          final keys = prefs.getKeys();
-          final all = <String, dynamic>{};
-          for (final key in keys) {
-            all[key] = prefs.get(key);
-          }
-          return developer.ServiceExtensionResponse.result(
-            jsonEncode({'status': 'Success', 'values': all}),
-          );
-
-        case 'get':
-          final key = params['key'];
-          if (key == null || key.isEmpty) {
-            return _errorResponse('missing key param');
-          }
-          final value = prefs.get(key);
-          return developer.ServiceExtensionResponse.result(
-            jsonEncode({
-              'status': 'Success',
-              'key': key,
-              'value': value,
-              'exists': value != null,
-            }),
-          );
-
-        case 'set':
-          final key = params['key'];
-          final raw = params['value'];
-          final type = params['type'] ?? 'string';
-          if (key == null || key.isEmpty) {
-            return _errorResponse('missing key param');
-          }
-          if (raw == null) return _errorResponse('missing value param');
-          switch (type) {
-            case 'bool':
-              await prefs.setBool(key, raw == 'true');
-            case 'int':
-              final n = int.tryParse(raw);
-              if (n == null) return _errorResponse('invalid int: $raw');
-              await prefs.setInt(key, n);
-            case 'double':
-              final d = double.tryParse(raw);
-              if (d == null) return _errorResponse('invalid double: $raw');
-              await prefs.setDouble(key, d);
-            default:
-              await prefs.setString(key, raw);
-          }
-          return developer.ServiceExtensionResponse.result(
-            jsonEncode({'status': 'Success', 'key': key, 'value': raw}),
-          );
-
-        case 'remove':
-          final key = params['key'];
-          if (key == null || key.isEmpty) {
-            return _errorResponse('missing key param');
-          }
-          await prefs.remove(key);
-          return developer.ServiceExtensionResponse.result(
-            jsonEncode({'status': 'Success', 'key': key}),
-          );
-
-        case 'clear':
-          await prefs.clear();
-          return developer.ServiceExtensionResponse.result(
-            jsonEncode({'status': 'Success'}),
-          );
-
-        default:
-          return _errorResponse(
-            'unknown action: $action. '
-            'Use get | getAll | set | remove | clear',
-          );
-      }
-    } catch (e) {
-      return _errorResponse('sharedPrefs failed: $e');
-    }
-  }
-
-  Future<developer.ServiceExtensionResponse> _handleClean(
-    String method,
-    Map<String, String> params,
-  ) async {
-    try {
-      final dirs = <Directory>[];
-
-      // Cache dir — getTemporaryDirectory() on iOS/Android
-      try {
-        dirs.add(await getTemporaryDirectory());
-      } catch (_) {}
-
-      // App support dir — persistent but non-user-facing storage
-      try {
-        dirs.add(await getApplicationSupportDirectory());
-      } catch (_) {}
-
-      // App documents dir — user-facing documents
-      try {
-        dirs.add(await getApplicationDocumentsDirectory());
-      } catch (_) {}
-
-      final cleaned = <String>[];
-      var totalFiles = 0;
-
-      for (final dir in dirs) {
-        if (!dir.existsSync()) continue;
-        final entities = dir.listSync(recursive: false);
-        for (final entity in entities) {
-          try {
-            if (entity is File) {
-              entity.deleteSync();
-              totalFiles++;
-            } else if (entity is Directory) {
-              entity.deleteSync(recursive: true);
-              totalFiles++;
-            }
-          } catch (_) {}
-        }
-        cleaned.add(dir.path);
-      }
-
-      return developer.ServiceExtensionResponse.result(
-        jsonEncode({
-          'status': 'Success',
-          'dirs': cleaned,
-          'deletedEntries': totalFiles,
-        }),
-      );
-    } catch (e) {
-      return _errorResponse('clean failed: $e');
-    }
-  }
-
-  /// Handles `ext.fdb.screenshot`.
-  ///
-  /// Renders the current Flutter surface to a PNG and returns it as a
-  /// base64-encoded string under the `screenshot` key.
-  ///
-  /// Used as a fallback capture backend on platforms that have no native
-  /// screenshot CLI (physical iOS, Windows, Linux Wayland).
-  Future<developer.ServiceExtensionResponse> _handleScreenshot(
-    String method,
-    Map<String, String> params,
-  ) async {
-    ui.Scene? scene;
-    ui.Image? image;
-    try {
-      final renderViews = WidgetsBinding.instance.renderViews;
-      if (renderViews.isEmpty) {
-        return _errorResponse('No render views available');
-      }
-
-      final view = renderViews.first;
-      final flutterView = view.flutterView;
-
-      // Ensure the frame is painted before capturing.
-      // ignore: invalid_use_of_protected_member
-      if (view.debugNeedsPaint || view.layer == null) {
-        WidgetsBinding.instance.scheduleFrame();
-        await WidgetsBinding.instance.endOfFrame;
-      }
-
-      // ignore: invalid_use_of_protected_member
-      final layer = view.layer;
-      if (layer == null) {
-        return _errorResponse('Render view layer is null');
-      }
-
-      // Capture at physical pixel resolution.
-      final size = flutterView.physicalSize;
-      final width = size.width.ceil();
-      final height = size.height.ceil();
-
-      if (width <= 0 || height <= 0) {
-        return _errorResponse('Invalid view size: ${width}x$height');
-      }
-
-      if (!layer.attached) {
-        return _errorResponse('Render view layer is detached');
-      }
-
-      final builder = ui.SceneBuilder();
-      layer.addToScene(builder);
-      scene = builder.build();
-      image = await scene.toImage(width, height);
-
-      final byteData = await image.toByteData(format: ui.ImageByteFormat.png);
-      if (byteData == null) {
-        return _errorResponse('Failed to encode image as PNG');
-      }
-
-      final base64Data = base64Encode(byteData.buffer.asUint8List());
-      return developer.ServiceExtensionResponse.result(
-        jsonEncode({'screenshot': base64Data}),
-      );
-    } catch (e) {
-      return _errorResponse('Screenshot failed: $e');
-    } finally {
-      scene?.dispose();
-      image?.dispose();
-    }
-  }
-
-  developer.ServiceExtensionResponse _errorResponse(String message) {
-    return developer.ServiceExtensionResponse.error(
-      -32000,
-      jsonEncode({'error': message}),
-    );
   }
 }
