+++ packages/fdb_helper/lib/src/handlers/handler_utils.dart
@@ -1,0 +1,9 @@
+import 'dart:convert';
+import 'dart:developer' as developer;
+
+developer.ServiceExtensionResponse errorResponse(String message) {
+  return developer.ServiceExtensionResponse.error(
+    -32000,
+    jsonEncode({'error': message}),
+  );
+}
